feat(search): refactor data client with updated service
