+++ src/services/SupabaseCachingService.ts
@@ -8,19 +8,15 @@ import { GetFractionsArgs } from "../graphql/schemas/args/fractionArgs.js";
 import { GetSalesArgs } from "../graphql/schemas/args/salesArgs.js";
 import { kyselyCaching } from "../client/kysely.js";
 import { supabaseCaching as supabaseClient } from "../client/supabase.js";
-import { expressionBuilder, Kysely, SqlBool } from "kysely";
 import { GetAllowlistRecordsArgs } from "../graphql/schemas/args/allowlistRecordArgs.js";
 import { singleton } from "tsyringe";
 import { BaseArgs } from "../graphql/schemas/args/baseArgs.js";
-import { SortOrder } from "../graphql/schemas/enums/sortEnums.js";
-import { buildWhereCondition } from "../graphql/schemas/utils/filters-kysely.js";
+import { BaseSupabaseService } from "./BaseSupabaseService.js";
 
 @singleton()
-export class SupabaseCachingService {
-  public readonly db: Kysely<CachingDatabase>;
-
+export class SupabaseCachingService extends BaseSupabaseService<CachingDatabase> {
   constructor() {
-    this.db = kyselyCaching;
+    super(kyselyCaching);
   }
 
   // Getters
@@ -227,86 +223,6 @@ export class SupabaseCachingService {
       default:
         throw new Error(`Table ${tableName.toString()} not found`);
     }
-  }
-
-  // Generalized query builder and handler of filter, sort, and pagination
-  handleGetCount<
-    DB extends CachingDatabase,
-    T extends keyof DB & string,
-    A extends object,
-  >(
-    tableName: T,
-    args: BaseArgs<A> & {
-      first?: number;
-      offset?: number;
-    },
-  ) {
-    let query = this.getCountQuery(tableName, args);
-
-    const { where } = args;
-    const eb = expressionBuilder(query);
-
-    if (where) {
-      query = this.applyWhereConditions(query, where, tableName, eb);
-    }
-
-    return query;
-  }
-
-  handleGetData<
-    DB extends CachingDatabase,
-    T extends keyof DB & string,
-    A extends object,
-  >(
-    tableName: T,
-    args: BaseArgs<A> & {
-      first?: number;
-      offset?: number;
-    },
-  ) {
-    let query = this.getDataQuery(tableName, args);
-    const { where, first, offset, sort } = args;
-    const eb = expressionBuilder(query);
-
-    if (where) {
-      query = this.applyWhereConditions(query, where, tableName, eb);
-    }
-
-    if (sort?.by) {
-      query = this.applySorting(query, sort.by);
-    }
-
-    if (first) query = query.limit(first);
-    if (offset) query = query.offset(offset);
-
-    return query;
-  }
-
-  private applyWhereConditions<T extends string>(
-    query: never,
-    where: never,
-    tableName: T,
-    eb: never,
-  ) {
-    const conditions = Object.entries(where)
-      .map(([column, value]) =>
-        buildWhereCondition(column, value, tableName, eb),
-      )
-      .filter(Boolean);
-
-    return conditions.reduce((q, condition) => {
-      return q.where(condition as SqlBool);
-    }, query);
-  }
-
-  private applySorting(query: never, sortBy: never) {
-    for (const [column, direction] of Object.entries(sortBy)) {
-      if (!column || !direction) continue;
-      const dir: "asc" | "desc" =
-        direction === SortOrder.ascending ? "asc" : "desc";
-      query = query.orderBy(column, dir);
-    }
-    return query;
   }
 
   getSalesForTokenIds(tokenIds: bigint[]) {
